+++ app/lib/src/pages/profile_page.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:talk_pilot/src/pages/profile_page/about_page.dart';
-import 'package:talk_pilot/src/pages/profile_page/faq_page.dart';
-import 'package:talk_pilot/src/pages/profile_page/help_page.dart';
 import 'package:talk_pilot/src/pages/profile_page/widgets/profile_card.dart';
 import 'package:talk_pilot/src/pages/profile_page/widgets/stats_card.dart';
+import 'package:talk_pilot/src/pages/profile_page/widgets/profile_drawer.dart';
 
 class ProfilePage extends StatefulWidget {
   const ProfilePage({super.key});
@@ -57,83 +54,7 @@ class _ProfilePageState extends State<ProfilePage> {
           ),
         ],
       ),
-      endDrawer: Drawer(
-        child: ListView(
-          padding: const EdgeInsets.symmetric(vertical: 30, horizontal: 10),
-          children: [
-            const ListTile(
-              title: Text(
-                '설정',
-                style: TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-              ),
-            ),
-            const Divider(),
-            ListTile(
-              leading: const Icon(Icons.info_outline),
-              title: const Text('About'),
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const AboutPage()),
-                );
-              },
-            ),
-            ListTile(
-              leading: const Icon(Icons.help_outline),
-              title: const Text('Help'),
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const HelpPage()),
-                );
-              },
-            ),
-            ListTile(
-              leading: const Icon(Icons.question_answer),
-              title: const Text('FAQ'),
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const FaqPage()),
-                );
-              },
-            ),
-            const Divider(),
-            ListTile(
-              leading: const Icon(Icons.logout, color: Colors.redAccent),
-              title: const Text(
-                '로그아웃',
-                style: TextStyle(color: Colors.redAccent),
-              ),
-              onTap: () {
-                showDialog(
-                  context: context,
-                  builder: (context) => AlertDialog(
-                    title: const Text('로그아웃 확인'),
-                    content: const Text('정말 로그아웃하시겠어요?'),
-                    actions: [
-                      TextButton(
-                        onPressed: () => Navigator.of(context).pop(),
-                        child: const Text('취소'),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          Navigator.of(context).pop();
-                          FirebaseAuth.instance.signOut();
-                        },
-                        child: const Text(
-                          '로그아웃',
-                          style: TextStyle(color: Colors.red),
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              },
-            ),
-          ],
-        ),
-      ),
+      endDrawer: const ProfileDrawer(),
       body: GestureDetector(
         onTap: () {
           if (isEditingNickname) {
@@ -196,35 +117,6 @@ class _ProfilePageState extends State<ProfilePage> {
           ],
         ),
       ),
-    );
-  }
-}
-
-/// 통계 항목 한 줄
-class StatRow extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final String value;
-
-  const StatRow({
-    super.key,
-    required this.icon,
-    required this.label,
-    required this.value,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Icon(icon, color: Colors.deepPurple),
-        const SizedBox(width: 10),
-        Expanded(child: Text(label, style: const TextStyle(fontSize: 16))),
-        Text(
-          value,
-          style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-        ),
-      ],
     );
   }
 }
